requirements.txt
langchain
langchain-community
langchain_openai
streamlit
tiktoken
python-dotenv
